Add files via upload
helpers.py gives us some useful things, though we still need to redesign the lookup function I think. or maybe we don't need it and we can just use SQL. Styles and layout are copied from Finance, so they should be fine for now just so we can view the page at least, and then we can update it when we wanna focus on design more. and then project.py should be pretty good through line 130 (register, login, and logout), and then I commented through the index page a bit. Buy and sell could be a pain, but we'll figure it out.